More fixes.

+++ firmware/src/apps/light_dimmer/light_dimmer.h
@@ -52,8 +52,8 @@ private:
 
     bool color_set = false;
 
-    uint16_t app_hue_position = 0;
-    uint16_t current_brightness = 0;
+    uint8_t app_hue_position = 0;
+    uint8_t current_brightness = 0;
     bool is_on = false;
 
     bool first_run = false;
